Revert "Merge branch 'main' into 로비구성_지민규"
This reverts commit 52f341f87ef509ba56cc8bf0ecffecb7cfd3333a, reversing
changes made to 82e838c532231c828f105979e06da5b67a5a80d7.

+++ Assets/Script/ScriptSO/UnitSO.cs
@@ -10,6 +10,6 @@ public class UnitSO : ScriptableObject
     public int atk;         // 공격력
     public int hp;          // 체력
     public int unitSpeed;   // 이동 속도
-    public float atkSpeed;    // 공격 속도
+    public int atkSpeed;    // 공격 속도
     public int unitScale;   // 유닛 크기
 }
